REFACTOR: move StructField to ClassDecl.StructDecl.StructField
