+++ base/src/main/java/org/aya/concrete/stmt/ClassDecl.java
@@ -3,8 +3,12 @@
 package org.aya.concrete.stmt;
 
 import kala.collection.immutable.ImmutableSeq;
+import kala.control.Option;
 import org.aya.concrete.Expr;
+import org.aya.concrete.Pattern;
 import org.aya.core.def.ClassDef;
+import org.aya.core.def.Def;
+import org.aya.core.def.FieldDef;
 import org.aya.core.def.StructDef;
 import org.aya.ref.DefVar;
 import org.aya.resolve.context.Context;
@@ -70,7 +74,7 @@ public sealed abstract class ClassDecl extends CommonDecl implements Decl.Result
   public static final class StructDecl extends ClassDecl {
     public final @NotNull DefVar<StructDef, StructDecl> ref;
     public @NotNull
-    final ImmutableSeq<TeleDecl.StructField> fields;
+    final ImmutableSeq<StructField> fields;
     public int ulift;
 
     public StructDecl(
@@ -80,7 +84,7 @@ public sealed abstract class ClassDecl extends CommonDecl implements Decl.Result
       @NotNull String name,
       @NotNull Expr result,
       // @NotNull ImmutableSeq<String> superClassNames,
-      @NotNull ImmutableSeq<TeleDecl.StructField> fields,
+      @NotNull ImmutableSeq<StructField> fields,
       @NotNull BindBlock bindBlock,
       @NotNull Decl.Personality personality
     ) {
@@ -93,5 +97,66 @@ public sealed abstract class ClassDecl extends CommonDecl implements Decl.Result
     @Override public @NotNull DefVar<StructDef, StructDecl> ref() {
       return ref;
     }
+
+      public static final class StructField extends CommonDecl implements Decl.Telescopic, Decl.Resulted {
+        public final @NotNull DefVar<FieldDef, StructField> ref;
+        public DefVar<StructDef, StructDecl> structRef;
+        public @NotNull ImmutableSeq<Pattern.Clause> clauses;
+        public @NotNull Expr result;
+        public @NotNull Option<Expr> body;
+        public final boolean coerce;
+
+        // will change after resolve
+        public @NotNull ImmutableSeq<Expr.Param> telescope;
+        public @Nullable Def.Signature signature;
+
+        public StructField(
+          @NotNull SourcePos sourcePos, @NotNull SourcePos entireSourcePos,
+          @Nullable OpInfo opInfo,
+          @NotNull String name,
+          @NotNull ImmutableSeq<Expr.Param> telescope,
+          @NotNull Expr result,
+          @NotNull Option<Expr> body,
+          @NotNull ImmutableSeq<Pattern.Clause> clauses,
+          boolean coerce,
+          @NotNull BindBlock bindBlock
+        ) {
+          super(sourcePos, entireSourcePos, Accessibility.Public, opInfo, bindBlock);
+          this.coerce = coerce;
+          this.result = result;
+          this.clauses = clauses;
+          this.body = body;
+          this.ref = DefVar.concrete(this, name);
+          this.telescope = telescope;
+        }
+
+        @Override public @NotNull DefVar<FieldDef, StructField> ref() {
+          return ref;
+        }
+
+        @Override public @NotNull ImmutableSeq<Expr.Param> telescope() {
+          return telescope;
+        }
+
+        @Override public void setTelescope(@NotNull ImmutableSeq<Expr.Param> telescope) {
+          this.telescope = telescope;
+        }
+
+        @Override public @Nullable Def.Signature signature() {
+          return signature;
+        }
+
+        @Override public void setSignature(Def.@Nullable Signature signature) {
+          this.signature = signature;
+        }
+
+        @Override public @NotNull Expr result() {
+          return result;
+        }
+
+        @Override public void setResult(@NotNull Expr result) {
+          this.result = result;
+        }
+      }
   }
 }
