+++ base/src/main/java/org/aya/concrete/stmt/TeleDecl.java
@@ -5,7 +5,6 @@ package org.aya.concrete.stmt;
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.MutableList;
 import kala.control.Either;
-import kala.control.Option;
 import org.aya.concrete.Expr;
 import org.aya.concrete.Pattern;
 import org.aya.core.def.*;
@@ -207,67 +206,6 @@ public sealed abstract class TeleDecl extends CommonDecl implements Decl.Telesco
 
     @Override public @NotNull DefVar<DataDef, DataDecl> ref() {
       return this.ref;
-    }
-  }
-
-  public static final class StructField extends CommonDecl implements Decl.Telescopic, Decl.Resulted {
-    public final @NotNull DefVar<FieldDef, TeleDecl.StructField> ref;
-    public DefVar<StructDef, ClassDecl.StructDecl> structRef;
-    public @NotNull ImmutableSeq<Pattern.Clause> clauses;
-    public @NotNull Expr result;
-    public @NotNull Option<Expr> body;
-    public final boolean coerce;
-
-    // will change after resolve
-    public @NotNull ImmutableSeq<Expr.Param> telescope;
-    public @Nullable Def.Signature signature;
-
-    public StructField(
-      @NotNull SourcePos sourcePos, @NotNull SourcePos entireSourcePos,
-      @Nullable OpInfo opInfo,
-      @NotNull String name,
-      @NotNull ImmutableSeq<Expr.Param> telescope,
-      @NotNull Expr result,
-      @NotNull Option<Expr> body,
-      @NotNull ImmutableSeq<Pattern.Clause> clauses,
-      boolean coerce,
-      @NotNull BindBlock bindBlock
-    ) {
-      super(sourcePos, entireSourcePos, Accessibility.Public, opInfo, bindBlock);
-      this.coerce = coerce;
-      this.result = result;
-      this.clauses = clauses;
-      this.body = body;
-      this.ref = DefVar.concrete(this, name);
-      this.telescope = telescope;
-    }
-
-    @Override public @NotNull DefVar<FieldDef, StructField> ref() {
-      return ref;
-    }
-
-    @Override public @NotNull ImmutableSeq<Expr.Param> telescope() {
-      return telescope;
-    }
-
-    @Override public void setTelescope(@NotNull ImmutableSeq<Expr.Param> telescope) {
-      this.telescope = telescope;
-    }
-
-    @Override public @Nullable Def.Signature signature() {
-      return signature;
-    }
-
-    @Override public void setSignature(Def.@Nullable Signature signature) {
-      this.signature = signature;
-    }
-
-    @Override public @NotNull Expr result() {
-      return result;
-    }
-
-    @Override public void setResult(@NotNull Expr result) {
-      this.result = result;
     }
   }
