+++ base/src/main/java/org/aya/core/def/FieldDef.java
@@ -5,7 +5,6 @@ package org.aya.core.def;
 import kala.collection.immutable.ImmutableSeq;
 import kala.control.Option;
 import org.aya.concrete.stmt.ClassDecl;
-import org.aya.concrete.stmt.TeleDecl;
 import org.aya.core.Matching;
 import org.aya.core.term.Term;
 import org.aya.ref.DefVar;
@@ -13,11 +12,11 @@ import org.jetbrains.annotations.NotNull;
 
 public final class FieldDef extends SubLevelDef {
   public final @NotNull DefVar<StructDef, ClassDecl.StructDecl> structRef;
-  public final @NotNull DefVar<FieldDef, TeleDecl.StructField> ref;
+  public final @NotNull DefVar<FieldDef, ClassDecl.StructDecl.StructField> ref;
   public final @NotNull Option<Term> body;
 
   public FieldDef(
-          @NotNull DefVar<StructDef, ClassDecl.StructDecl> structRef, @NotNull DefVar<FieldDef, TeleDecl.StructField> ref,
+          @NotNull DefVar<StructDef, ClassDecl.StructDecl> structRef, @NotNull DefVar<FieldDef, ClassDecl.StructDecl.StructField> ref,
           @NotNull ImmutableSeq<Term.Param> ownerTele, @NotNull ImmutableSeq<Term.Param> selfTele,
           @NotNull Term result, @NotNull ImmutableSeq<Matching> clauses, @NotNull Option<Term> body, boolean coerce
   ) {
@@ -36,7 +35,7 @@ public final class FieldDef extends SubLevelDef {
     return selfTele;
   }
 
-  public @NotNull DefVar<FieldDef, TeleDecl.StructField> ref() {
+  public @NotNull DefVar<FieldDef, ClassDecl.StructDecl.StructField> ref() {
     return ref;
   }
 }
