+++ base/src/main/java/org/aya/core/term/RefTerm.java
@@ -3,7 +3,6 @@
 package org.aya.core.term;
 
 import org.aya.concrete.stmt.ClassDecl;
-import org.aya.concrete.stmt.TeleDecl;
 import org.aya.core.def.FieldDef;
 import org.aya.core.def.StructDef;
 import org.aya.core.pat.Pat;
@@ -20,7 +19,7 @@ public record RefTerm(@NotNull LocalVar var, int lift) implements Term {
     return visitor.visitRef(this, p);
   }
 
-  public record Field(@NotNull DefVar<FieldDef, TeleDecl.StructField> ref, int lift) implements Term {
+  public record Field(@NotNull DefVar<FieldDef, ClassDecl.StructDecl.StructField> ref, int lift) implements Term {
     @Override public <P, R> R doAccept(@NotNull Visitor<P, R> visitor, P p) {
       return visitor.visitFieldRef(this, p);
     }
